BJS2-78419: fix
