--- src/main/java/faang/school/urlshortenerservice/andreev/controller/HashController.java
@@ -1,28 +1,0 @@
-package faang.school.urlshortenerservice.andreev.controller;
-
-import faang.school.urlshortenerservice.andreev.dto.UrlRequestDto;
-import faang.school.urlshortenerservice.andreev.dto.UrlResponseDto;
-import faang.school.urlshortenerservice.andreev.service.hash.HashService;
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/api/v1/urls")
-@RequiredArgsConstructor
-public class HashController {
-    private final HashService hashService;
-
-    @PostMapping
-    public UrlResponseDto createShortUrl(@Valid @NotNull @NotBlank @RequestBody UrlRequestDto urlRequestDto) {
-        return hashService.createShortUrl(urlRequestDto);
-    }
-
-    @GetMapping("/{hash}")
-    public ResponseEntity<String> redirectToOriginalUrl(@Valid @NotNull @NotBlank @PathVariable String hash) {
-        return hashService.redirectToOriginalUrl(hash);
-    }
-}
