--- src/main/java/faang/school/urlshortenerservice/andreev/entity/Hash.java
@@ -1,24 +1,0 @@
-package faang.school.urlshortenerservice.andreev.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "hash")
-public class Hash {
-    @Id
-    @Column(name = "hash", nullable = false, unique = true, length = 6)
-    private String hash;
-}
