--- src/main/java/faang/school/urlshortenerservice/andreev/entity/Url.java
@@ -1,32 +1,0 @@
-package faang.school.urlshortenerservice.andreev.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Table(name = "url")
-public class Url {
-    @Id
-    @Column(name = "hash", nullable = false, unique = true, length = 6)
-    private String hash;
-
-    @Column(name = "url", nullable = false)
-    private String url;
-
-    @CreationTimestamp
-    @Column(name = "created_at", nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-}
