--- src/main/java/faang/school/urlshortenerservice/andreev/exception/ErrorMessage.java
@@ -1,7 +1,0 @@
-package faang.school.urlshortenerservice.andreev.exception;
-
-public class ErrorMessage {
-    public static final String HASH_GENERATION_FAILED = "Failed to generate hashes";
-    public static final String URL_NOT_FOUND_BY_HASH = "Url not found by hash: %s";
-    public static final String INVALID_URL = "The URL %s is not the correct format";
-}
