--- src/main/java/faang/school/urlshortenerservice/andreev/exception/HashGenerationException.java
@@ -1,7 +1,0 @@
-package faang.school.urlshortenerservice.andreev.exception;
-
-public class HashGenerationException extends RuntimeException {
-    public HashGenerationException(String message) {
-        super(message);
-    }
-}
