--- src/main/java/faang/school/urlshortenerservice/andreev/exception/InvalidUrlException.java
@@ -1,7 +1,0 @@
-package faang.school.urlshortenerservice.andreev.exception;
-
-public class InvalidUrlException extends RuntimeException {
-    public InvalidUrlException(String message) {
-        super(message);
-    }
-}
