--- src/main/java/faang/school/urlshortenerservice/andreev/exception/UrlNotFound.java
@@ -1,7 +1,0 @@
-package faang.school.urlshortenerservice.andreev.exception;
-
-public class UrlNotFound extends RuntimeException {
-    public UrlNotFound(String message) {
-        super(message);
-    }
-}
