+++ src/main/java/faang/school/urlshortenerservice/dto/UrlRequestDto.java
@@ -1,0 +1,12 @@
+package faang.school.urlshortenerservice.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class UrlRequestDto {
+    private String url;
+}
