+++ src/main/java/faang/school/urlshortenerservice/dto/UrlResponseDto.java
@@ -1,0 +1,20 @@
+package faang.school.urlshortenerservice.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.io.Serial;
+import java.io.Serializable;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class UrlResponseDto implements Serializable {
+    @Serial
+    private static final long serialVersionUID = 1L;
+
+    private String hash;
+
+    private String url;
+}
