+++ src/main/java/faang/school/urlshortenerservice/entity/Hash.java
@@ -1,0 +1,22 @@
+package faang.school.urlshortenerservice.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "hash")
+public class Hash {
+    @Id
+    @Column(name = "hash", nullable = false, unique = true, length = 6)
+    private String hash;
+}
