+++ src/main/java/faang/school/urlshortenerservice/entity/Url.java
@@ -1,0 +1,32 @@
+package faang.school.urlshortenerservice.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Table(name = "url")
+public class Url {
+    @Id
+    @Column(name = "hash", nullable = false, unique = true, length = 6)
+    private String hash;
+
+    @Column(name = "url", nullable = false, length = 2048)
+    private String url;
+
+    @CreationTimestamp
+    @Column(name = "created_at", nullable = false, updatable = false)
+    private LocalDateTime createdAt;
+}
