+++ src/main/java/faang/school/urlshortenerservice/exception/ErrorMessage.java
@@ -1,0 +1,7 @@
+package faang.school.urlshortenerservice.exception;
+
+public class ErrorMessage {
+    public static final String HASH_GENERATION_FAILED = "Failed to generate hashes";
+    public static final String URL_NOT_FOUND_BY_HASH = "Url not found by hash: %s";
+    public static final String INVALID_URL = "The URL %s is not the correct format";
+}
