+++ src/main/java/faang/school/urlshortenerservice/exception/HashGenerationException.java
@@ -1,0 +1,7 @@
+package faang.school.urlshortenerservice.exception;
+
+public class HashGenerationException extends RuntimeException {
+    public HashGenerationException(String message) {
+        super(message);
+    }
+}
