+++ src/main/java/faang/school/urlshortenerservice/exception/InvalidUrlException.java
@@ -1,0 +1,7 @@
+package faang.school.urlshortenerservice.exception;
+
+public class InvalidUrlException extends RuntimeException {
+    public InvalidUrlException(String message) {
+        super(message);
+    }
+}
