+++ src/main/java/faang/school/urlshortenerservice/exception/UrlNotFound.java
@@ -1,0 +1,7 @@
+package faang.school.urlshortenerservice.exception;
+
+public class UrlNotFound extends RuntimeException {
+    public UrlNotFound(String message) {
+        super(message);
+    }
+}
