+++ src/main/java/faang/school/urlshortenerservice/service/hash/HashService.java
@@ -1,0 +1,11 @@
+package faang.school.urlshortenerservice.service.hash;
+
+import faang.school.urlshortenerservice.dto.UrlRequestDto;
+import faang.school.urlshortenerservice.dto.UrlResponseDto;
+import org.springframework.http.ResponseEntity;
+
+public interface HashService {
+    public ResponseEntity<String> redirectToOriginalUrl(String hash);
+
+    public UrlResponseDto createShortUrl(UrlRequestDto url);
+}
