+++ src/main/resources/db/changelog/changeset/url_v000__create_hash_url.sql
@@ -11,10 +11,10 @@ CREATE TABLE IF NOT EXISTS  hash
 CREATE TABLE IF NOT EXISTS url
 (
     hash       VARCHAR(6) NOT NULL,
-    url        TEXT       NOT NULL,
+    url        VARCHAR(2048)       NOT NULL,
     created_at TIMESTAMP  NOT NULL DEFAULT CURRENT_TIMESTAMP,
     CONSTRAINT url_pk PRIMARY KEY (hash)
 );
 
 CREATE INDEX IF NOT EXISTS  url_created_at_idx ON url(created_at);
-CREATE INDEX IF NOT EXISTS  url_url_idx ON url(url);
+CREATE INDEX IF NOT EXISTS  url_url_idx ON url USING hash (url);
